Arel::Nodes::RangeFunction
